Refactor: encapsulate SpiceConnection
Create a GObject-derived SpiceConnection class that encapsulates a
connection to the Spice server. Remove the SpiceWindow class as only a
SpiceDisplay is needed. Encapsulate the SpiceDisplay instance inside the
connection instance. Create a singleton SpiceConnection instance named
mainconn, and get the global_display instance through it. Properly
destroy the SpiceConnection and SpiceDisplay instances on disconnect.
Remove many unused or unreachable code.